change physicalID from 32bit to 64bit

+++ Motion_controller/src/fst_io_device/include/fst_io_device.h
@@ -72,16 +72,16 @@ public:
     IODeviceInfo getDeviceInfo(void);
     IODevicePortValues getDeviceValues(void);
 
-    ErrorCode getDiValue(uint8_t port_offset, uint8_t &value);
-    ErrorCode getDoValue(uint8_t port_offset, uint8_t &value);
-    ErrorCode getRiValue(uint8_t port_offset, uint8_t &value);
-    ErrorCode getRoValue(uint8_t port_offset, uint8_t &value);
-    ErrorCode getUiValue(uint8_t port_offset, uint8_t &value);
-    ErrorCode getUoValue(uint8_t port_offset, uint8_t &value);
+    ErrorCode getDiValue(uint32_t port_offset, uint8_t &value);
+    ErrorCode getDoValue(uint32_t port_offset, uint8_t &value);
+    ErrorCode getRiValue(uint32_t port_offset, uint8_t &value);
+    ErrorCode getRoValue(uint32_t port_offset, uint8_t &value);
+    ErrorCode getUiValue(uint32_t port_offset, uint8_t &value);
+    ErrorCode getUoValue(uint32_t port_offset, uint8_t &value);
 
-    ErrorCode setDoValue(uint8_t port_offset, uint8_t value);
-    ErrorCode setRoValue(uint8_t port_offset, uint8_t value);
-    ErrorCode setUoValue(uint8_t port_offset, uint8_t value);
+    ErrorCode setDoValue(uint32_t port_offset, uint8_t value);
+    ErrorCode setRoValue(uint32_t port_offset, uint8_t value);
+    ErrorCode setUoValue(uint32_t port_offset, uint8_t value);
 
     ErrorCode updateDeviceData(void);
